NavigationHelper created.
